% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imago-package.R, R/imago.R
\docType{package}
\name{imago}
\alias{imago}
\alias{imago-package}
\title{Hacky World Map GeoJSON Based on the Imago Projection}
\usage{
imago(...)
}
\arguments{
\item{...}{passed on to \code{\link[sf:st_read]{sf::st_read()}}}
}
\description{
In 2017, Justin Kunimune reverse engineerd
(\url{https://kunimune.home.blog/2017/11/23/the-secrets-of-the-authagraph-revealed/})
and open sourced the proprietary AuthaGraph (http://www.authagraph.com/top/?lang=en)
projection. Tools are provided to use this map in R.

The resultant simplefeatures object has a \code{iso3c} column
so you can bind values. A most-certainly non-iso3c value of
\code{lakes} is used to identify inland bodies of water.
}
\examples{
plot(imago())
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/imago}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/imago/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
